requirements.txt
filetype
Pillow
urllib3
munch
pygifsicle